add different prior(\beta) to word distribution

+++ src/structures/_ChildDoc.java
@@ -3,8 +3,8 @@ package structures;
 import utils.Utils;
 
 public class _ChildDoc extends _Doc {
-	public int[][] m_xTopicSstat;//joint assignment of <x,z>: 0 from general, 1 from specific
-	public int[] m_xSstat; // sufficient statistics for x
+	public double[][] m_xTopicSstat;//joint assignment of <x,z>: 0 from general, 1 from specific
+	public double[] m_xSstat; // sufficient statistics for x
 
 	public double[][] m_xTopics; // proportion of topics (0 from general, 1 from specific)
 	public double[] m_xProportion; // proportion of x
@@ -31,9 +31,9 @@ public class _ChildDoc extends _Doc {
 	}
 
 	public void createXSpace(int k, int gammaSize) {
-		m_xTopicSstat = new int[gammaSize][k];
+		m_xTopicSstat = new double[gammaSize][k];
 		m_xTopics = new double[gammaSize][k];
-		m_xSstat = new int[gammaSize];
+		m_xSstat = new double[gammaSize];
 		m_xProportion = new double[gammaSize];
 	}
